requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
platformdirs==3.10.0
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.2
